Move Isoformatter into test helper, add date/time tests
